jednoduche upravy

+++ map/change_route_status.php
@@ -9,21 +9,26 @@ session_start();
 
 //var_dump($_POST[]);
 
-if (!empty($_POST['id']))
+$bu = $link->query(sprintf("SELECT * FROM actives WHERE user_id = %d", $_SESSION['id']));
+
+if ($bu->num_rows > 0)
 {
-	//$q = $link->query(sprintf("SELECT * FROM actives WHERE user_id = %d AND route_id = %d", $_SESSION['id'], $_POST['id']));
-    $q = $link->query(sprintf("SELECT * FROM actives WHERE route_id = %d", $_POST['id']));
+    echo "uz mas aktivnu trasu";
 
-    if ($q->num_rows > 0)
-	{
-        //$q = $link->query(sprintf("DELETE FROM actives WHERE user_id = %d AND route_id = %d", $_SESSION['id'], $_POST['id']));
-        $q = $link->query(sprintf("DELETE FROM actives WHERE route_id = %d", $_POST['id']));
+} else {
+
+    if (!empty($_POST['id'])) {
+        $q = $link->query(sprintf("SELECT * FROM actives WHERE user_id = %d AND route_id = %d", $_SESSION['id'], $_POST['id']));
+//    $q = $link->query(sprintf("SELECT * FROM actives WHERE route_id = %d", $_POST['id']));
+
+        if ($q->num_rows > 0) {
+            $q = $link->query(sprintf("DELETE FROM actives WHERE user_id = %d AND route_id = %d", $_SESSION['id'], $_POST['id']));
+//        $q = $link->query(sprintf("DELETE FROM actives WHERE route_id = %d", $_POST['id']));
+        } else {
+            $q = $link->query(sprintf("INSERT INTO actives (user_id, route_id ) VALUES (%d,%d)", $_SESSION['id'], $_POST['id']));
+//        $q = $link->query(sprintf("INSERT INTO actives ( route_id ) VALUES (%d)", $_POST['id']));
+        }
+
     }
-	else
-	{
-//        $q = $link->query(sprintf("INSERT INTO actives (user_id, route_id ) VALUES (%d,%d)", $_SESSION['id'], $_POST['id']));
-        $q = $link->query(sprintf("INSERT INTO actives ( route_id ) VALUES (%d)", $_POST['id']));
-	}
-
 }
